refactor: Move static method 'of' to parent class towards to future additions
